chore(secretsmanager): migrate tests to jest (#10768)
No tests or expectations added or removed.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*
