+++ packages/@aws-cdk/aws-secretsmanager/.gitignore
@@ -12,5 +12,6 @@ coverage
 dist
 tsconfig.json
 !.eslintrc.js
+!jest.config.js
 
 junit.xml
