+++ packages/@aws-cdk/aws-secretsmanager/.npmignore
@@ -22,6 +22,7 @@ dist
 
 tsconfig.json
 .eslintrc.js
+jest.config.js
 
 # exclude cdk artifacts
 **/cdk.out
