+++ packages/@aws-cdk/aws-secretsmanager/jest.config.js
@@ -1,0 +1,2 @@
+const baseConfig = require('cdk-build-tools/config/jest.config');
+module.exports = baseConfig;
